pyproject.toml
[project]
name = "dbx-pipeline"
version = "0.1.0"
description = ""
authors = [
    {name = "omarkhano2403@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "databricks-dlt (>=0.3.0,<0.4.0)"
]

[tool.poetry]
packages = [{include="dbx_data_pipeline", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
